+++ sgab-view-web/web/core/pessoa/alterar.jsp
@@ -9,7 +9,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <title>SGAB | Resultado de Pessoa</title>
     <meta content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no" name="viewport">
-    <link rel="stylesheet" href="../../css/styles.css">
+    <link rel="stylesheet" href="/sgab/css/styles.css">
     <style>
         #nome, #email, #senha {
             background: #fff url(imgs/editar.png) no-repeat 95% center;
@@ -31,7 +31,7 @@
         </div>
         <div id="logo-topo">
             <a href="http://www.cefetmg.br/" target="_blank" title="Centro Federal de Educação Tecnológica de Minas Gerais">
-                <img src="../../images/logo_topo.png" alt="Centro Federal de Educação Tecnológica de Minas Gerais">
+                <img src="/sgab/images/logo_topo.png" alt="Centro Federal de Educação Tecnológica de Minas Gerais">
             </a>
         </div>
     </header>
@@ -46,7 +46,7 @@
                     <input type="hidden" name="acao" value="alterar">
                     <div>
                         <label for="cpf">Código</label>
-                        <input type="number" id="id" name="id" placeholder="ID" value="<%=pessoa.getId()%>" readonly>
+                        <input type="number" id="pessoaId" name="pessoaId" placeholder="ID" value="<%=pessoa.getId()%>" readonly>
                     </div>
                     <div>
                         <label for="cpf">CPF</label>
@@ -87,7 +87,8 @@
     <footer>
         <p>SGAB - Sistema de Gestão de Acervo Bibliográfico</p>
     </footer>
-    <script src="../../js/validacao.js"></script>
+    <script type="text/javascript" language="JavaScript" src="/sgab/js/validacao.js"></script>
+    <script type="text/javascript" language="JavaScript" src="/sgab/js/helper.js"></script>
     <script>
         let senha2DivEl = document.querySelector("#senha2-div");
