+++ sgab-view-web/web/core/pessoa/inserir.jsp
@@ -7,7 +7,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <title>SGAB | Cadastro de Pessoa</title>
     <meta content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no" name="viewport">
-    <link rel="stylesheet" href="../../css/styles.css">
+    <link rel="stylesheet" href="/sgab/css/styles.css">
 </head>
 <body>
     <header>
@@ -18,7 +18,7 @@
         </div>
         <div id="logo-topo">
             <a href="http://www.cefetmg.br/" target="_blank" title="Centro Federal de Educação Tecnológica de Minas Gerais">
-                <img src="../../images/logo_topo.png" alt="Centro Federal de Educação Tecnológica de Minas Gerais">
+                <img src="/sgab/images/logo_topo.png" alt="Centro Federal de Educação Tecnológica de Minas Gerais">
             </a>
         </div>
     </header>
@@ -72,6 +72,7 @@
     <footer>
         <p>SGAB - Sistema de Gestão de Acervo Bibliográfico</p>
     </footer>
-    <script src="../../js/validacao.js"></script>
+    <script type="text/javascript" language="JavaScript" src="/sgab/js/validacao.js"></script>
+    <script type="text/javascript" language="JavaScript" src="/sgab/js/helper.js"></script>
 </body>
 </html>
