+++ sgab-view-web/web/core/pessoa/pesquisar.jsp
@@ -7,7 +7,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <title>SGAB | Pesquisa de Pessoa</title>
     <meta content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no" name="viewport">
-    <link rel="stylesheet" href="../../css/styles.css">
+    <link rel="stylesheet" href="/sgab/css/styles.css">
 </head>
 <body>
     <header>
@@ -18,7 +18,7 @@
         </div>
         <div id="logo-topo">
             <a href="http://www.cefetmg.br/" target="_blank" title="Centro Federal de Educação Tecnológica de Minas Gerais">
-                <img src="../../images/logo_topo.png" alt="Centro Federal de Educação Tecnológica de Minas Gerais">
+                <img src="/sgab/images/logo_topo.png" alt="Centro Federal de Educação Tecnológica de Minas Gerais">
             </a>
         </div>
     </header>
@@ -30,7 +30,7 @@
             <div id="caixa-form">
                 <form name="perquisarPessoa" method="post">
                     <input type="hidden" name="table" value="Pessoa">
-                    <input type="hidden" name="acao" value="pesquisar">
+                    <input type="hidden" name="acao" value="pesquisarPorLogin">
                     <div>
                         <label for="login">Informe o Login <span class="input-obrigatorio">*</span></label>
                         <input type="number" id="login" name="login" placeholder="Login" required>
@@ -46,6 +46,7 @@
     <footer>
         <p>SGAB - Sistema de Gestão de Acervo Bibliográfico</p>
     </footer>
-   <script src="../../js/validacao.js"></script> 
+    <script type="text/javascript" language="JavaScript" src="/sgab/js/validacao.js"></script>
+    <script type="text/javascript" language="JavaScript" src="/sgab/js/helper.js"></script>
 </body>
 </html>
