+++ sgab-view-web/web/js/validacao.js
@@ -10,15 +10,16 @@ function validacaoView(inputEl, regexEl, smallEl, labelEl, inputObrigatorioEl) {
         labelEl.style.color = "#dd4b39";
         inputEl.style.borderColor = "#dd4b39";
         inputObrigatorioEl.style.visibility = "visible";
+        inputEl.focus();
     }
 }
 
 let loginEl = document.querySelector("#login");
-cpfEl.addEventListener("change", () => {
+loginEl.addEventListener("change", () => {
 
-    let smallEl = document.querySelector("#cpf ~ small");
-    let labelEl = document.querySelector("label[for=cpf]");
-    let inputObrigatorioEl = document.querySelector("label[for=cpf] .input-obrigatorio");
+    let smallEl = document.querySelector("#login ~ small");
+    let labelEl = document.querySelector("label[for=login]");
+    let inputObrigatorioEl = document.querySelector("label[for=login] .input-obrigatorio");
     
     if(loginEl == null){
         smallEl.classList.add("invalido");
@@ -82,75 +83,3 @@ senha2El.addEventListener("change", () => {
 
     validacaoView(senha2El, regexEl, smallEl, labelEl, inputObrigatorioEl);
 });
-
-function validarPessoa(frm) {
-    var result = false;
-    
-    if (frm.cpf.value == "") {
-        alert("Informar o cpf!");
-        frm.cpf.focus();
-    } else if(frm.login.value === ""){
-        alert("Informar o Login!");
-        frm.login.focus();
-    } else if (frm.nome.value === "") {
-        alert("Informar o nome!");
-        frm.nome.focus();
-    } else if (frm.senha.value === "") {
-        alert("Informar a senha!");
-        frm.senha.focus();
-    } else if (frm.email.value === "") {
-        alert("Informar o email!");
-        frm.email.focus();
-    } 
-    else
-        result = true;
-        
-    return result;
-
-}
-
-
-function gravar(frm) {
-    var table = frm.table.value;
-
-    if (table === "Pessoa") {
-        if (validarPessoa(frm)) {
-            if (frm.acao.value === "alterar")
-                caminhourl = "/sgab/main?acao=PessoaGravarAlteracao";
-            else if (frm.acao.value === "gravar")
-                caminhourl = "/sgab/main?acao=PessoaGravarInsercao";
-        }
-    }
-
-    frm.action = caminhourl;
-    frm.submit();
-}
-
-function pesquisar(frm){
-    var table = frm.table.value;
-
-    if (table === "Pessoa") {
-        if (frm.login.value == "") {
-            alert("Informar o login!");
-            frm.cpf.focus();
-        } else {
-            frm.action = "/sgab/main?acao=PessoaPesquisar&PessoaLogin=" + frm.login.value;            
-            frm.submit();
-        }
-
-
-    }
-}
-
-function excluir(cpf, frm) {
-    var table = frm.table.value;
-
-    if (table === "Pessoa") {
-
-        if (confirm('Deseja excluir o Usuário com Id = ' + cpf + '?')) {
-            frm.cpf.value = cpf;
-            frm.action = "/sgab/main?acao=PessoaExcluir";            
-            frm.submit();
-        }
-    }
-}
